Move feature to folder push-notifications. Use feature toggle
