--- src/notifications/index.ts
@@ -1,1 +1,0 @@
-export * from './use-push-notifications';
