+++ src/push-notifications/index.ts
@@ -2,3 +2,4 @@ export {
   usePushNotificationsEnabledDebugOverride,
   usePushNotificationsEnabled,
 } from './use-push-notifications-enabled';
+export * from './use-push-notifications';
