+++ src/push-notifications/use-push-notifications.tsx
@@ -1,6 +1,6 @@
 import {useCallback, useState} from 'react';
-import {useRegister} from '@atb/notifications/use-register';
-import {useConfig} from '@atb/notifications/use-config';
+import {useRegister} from '@atb/push-notifications/use-register';
+import {useConfig} from '@atb/push-notifications/use-config';
 import {PermissionsAndroid, Platform} from 'react-native';
 import messaging, {
   FirebaseMessagingTypes,
