+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_ProfileStack/Profile_RootScreen.tsx
@@ -48,6 +48,7 @@ import {
 import {RootStackParamList} from '@atb/stacks-hierarchy';
 import {InfoTag} from '@atb/components/info-tag';
 import {ClickableCopy} from './components/ClickableCopy';
+import {usePushNotificationsEnabled} from '@atb/push-notifications';
 
 const buildNumber = getBuildNumber();
 const version = getVersion();
@@ -90,6 +91,7 @@ export const Profile_RootScreen = ({navigation}: ProfileProps) => {
 
   const phoneNumber = parsePhoneNumber(authPhoneNumber ?? '');
   const {enable_vipps_login} = useRemoteConfig();
+  const isPushNotificationsEnabled = usePushNotificationsEnabled();
 
   const {open: openBottomSheet, close: closeBottomSheet} = useBottomSheet();
   async function selectFavourites() {
@@ -292,14 +294,16 @@ export const Profile_RootScreen = ({navigation}: ProfileProps) => {
             onPress={() => navigation.navigate('Profile_LanguageScreen')}
             testID="languageButton"
           />
-          <LinkSectionItem
-            text={t(
-              ProfileTexts.sections.settings.linkSectionItems.notifications
-                .label,
-            )}
-            onPress={() => navigation.navigate('Profile_NotificationsScreen')}
-            testID="notificationsButton"
-          />
+          {isPushNotificationsEnabled && (
+            <LinkSectionItem
+              text={t(
+                ProfileTexts.sections.settings.linkSectionItems.notifications
+                  .label,
+              )}
+              onPress={() => navigation.navigate('Profile_NotificationsScreen')}
+              testID="notificationsButton"
+            />
+          )}
         </Section>
         <Section withPadding>
           <GenericSectionItem>
